removing some disps and explaining the previous commit
MArcelo adviced me not to fill empty cells with [0.5 0.5 0.5] but rather
fill the distance matrix. I did it and tested it. When I was filling
empty cells, the code was finding the red spot, but now it cannot find
it anymore. Why? Because of min(). Such function is returning the
artifical line [0.5 0.5] because it is the shortest distance. Now I need
to work on the variable detectado.